Fixed Mobile Errors and Optimized
Fixed errors on mobile and optimized image displaying using NextImage component
