+++ components/nav-bar/nav-bar.tsx
@@ -7,7 +7,7 @@ export default function NavBar({ showOnScroll }: { showOnScroll: boolean }) {
 	const links: NavBarLinkType[] = [
 		{ href: "/", children: "Home" },
 		{ href: "/sponsor-us", children: "Sponsor Us" },
-		{ href: "/MeetTheTeamPage", children: "Team" },
+		{ href: "/Team", children: "Team" },
 	];
 
 	const [showNav, setShowNav] = useState(false);
